pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ha-addon-scheiber"
version = "1.3.2"
description = "Home Assistant addon for Scheiber CAN bus devices"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "eburi"}
]

dependencies = [
    "python-can==4.3.1",
    "paho-mqtt==2.1.0",
    "PyYAML==6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "poethepoet>=0.24.0",
]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true
extend_skip = [".venv", "__pycache__"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

# Poe the Poet configuration
[tool.poe]
# Use the virtualenv Python interpreter
executor = { type = "virtualenv", location = "scheiber/src/.venv" }

# Poe the Poet task definitions
[tool.poe.tasks]

[tool.poe.tasks.install]
help = "Install production dependencies"
cmd = "pip install -r scheiber/src/requirements.txt"

[tool.poe.tasks.install-dev]
help = "Install development dependencies"
cmd = "pip install pytest black isort poethepoet"

[tool.poe.tasks.test]
help = "Run all tests"
shell = "python -m pytest tests/ -v"

[tool.poe.tasks.test-coverage]
help = "Run tests with coverage report"
shell = "python -m pytest tests/ -v --cov=scheiber/src --cov-report=term-missing"

[tool.poe.tasks.format]
help = "Format code with black and isort"
sequence = [
    {shell = "python -m black scheiber/src/ tests/"},
    {shell = "python -m isort scheiber/src/ tests/ --profile black"},
]

[tool.poe.tasks.check-format]
help = "Check code formatting without making changes"
sequence = [
    {shell = "python -m black --check scheiber/src/ tests/"},
    {shell = "python -m isort --check-only scheiber/src/ tests/ --profile black"},
]

[tool.poe.tasks.imports]
help = "Organize imports in all Python files"
shell = "python -m isort scheiber/src/ tests/ --profile black"

[tool.poe.tasks.check-imports]
help = "Check import organization without making changes"
shell = "python -m isort --check-only --diff scheiber/src/ tests/ --profile black"

[tool.poe.tasks.lint]
help = "Run linter (currently checks format and imports)"
sequence = [
    {shell = "python -m black --check scheiber/src/ tests/"},
    {shell = "python -m isort --check-only scheiber/src/ tests/ --profile black"},
]

[tool.poe.tasks.check]
help = "Run all checks (format, imports, tests)"
sequence = ["check-format", "check-imports", "test"]

[tool.poe.tasks.clean]
help = "Remove Python cache files"
shell = """
find . -type d -name __pycache__ -exec rm -rf {} + 2>/dev/null || true
find . -type f -name "*.pyc" -delete 2>/dev/null || true
find . -type d -name "*.egg-info" -exec rm -rf {} + 2>/dev/null || true
"""

[tool.poe.tasks.run-bridge]
help = "Run MQTT bridge with default settings"
cmd = "python scheiber/src/mqtt_bridge.py --log-level info"

[tool.poe.tasks.run-bridge-debug]
help = "Run MQTT bridge with debug logging"
cmd = "python scheiber/src/mqtt_bridge.py --log-level debug"

[tool.poe.tasks.canlistener]
help = "Run CAN listener tool (usage: poe canlistener can1)"
cmd = "python scheiber/src/tools/canlistener.py"

[tool.poe.tasks.analyser]
help = "Run interactive CAN analyzer (usage: poe analyser -i can1)"
cmd = "python scheiber/src/tools/analyser.py"

[tool.poe.tasks.bloc9-status]
help = "Run Bloc9 status analyzer (usage: poe bloc9-status can1 [bloc9_id])"
cmd = "python scheiber/src/tools/bloc9_status.py"
